==> 02-additional-data.ipynb <==
# Purpose

Download and transform open source data that can be used to predict travel times.
Each data set needs to be able to join onto the original dataset using LSOA name and have data for years 2014 and 2015.

**Ideas**

* [Number of schools from DfE](https://www.gov.uk/government/collections/statistics-school-and-pupil-numbers)
* Number of GPs and hosptials from [here](http://hefs.hscic.gov.uk/DataFiles.asp). [Postcodes](https://digital.nhs.uk/organisation-data-service/data-downloads/postcode-files).[NHS digital data downloads](https://digital.nhs.uk/organisation-data-service/data-downloads). 
* Size of road network from [DfT stats page](https://www.gov.uk/government/collections/road-network-size-and-condition)
* Public transport measures [from DfT](https://www.gov.uk/government/organisations/department-for-transport/about/statistics) (number of bus/train stations, number of bus/train journeys)
* [Population from ONS](https://www.ons.gov.uk/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/lowersuperoutputareamidyearpopulationestimates) (include demographics such as children, students, pensioners and students)
* [ONS rural-urban classification](http://geoportal.statistics.gov.uk/datasets?q=RUC_2011_LSOA&sort=name)


The code below downloads(if download=True) the data sets and cleans them and joins them onto the travel time statistics. I could not find any data for rail. All data was available for 2014 and 2015, some for 2016.
Data was usually at local authority (LA) level and so mapped to a LSOA.

import zipfile
import os
import requests 
import shutil
import pyexcel
import pandas as pd
import numpy as np

pd.set_option('display.max_columns', 30)

# Create folder to store data in, if it doesn't already exist
if os.path.isdir('open-data') == False:
    os.mkdir('open-data')
    
def download_file(url, filepath_to_save_to):
    r = requests.get(url, stream=True)
    if r.status_code == 200:
        with open(filepath_to_save_to, 'wb') as out_file:
            shutil.copyfileobj(r.raw, out_file)
    else: print('Download failed')

def download_unzip(url, file_to_extract, save_as):
        filepath = 'open-data/population-density.zip'
        download_file(url, 'open-data/zipfile.zip')
        with zipfile.ZipFile('open-data/zipfile.zip') as myzip:
            myzip.extract(file_to_extract,
                         path='open-data')
        os.rename('open-data/' + file_to_extract, 'open-data/' + save_as)
        os.remove('open-data/zipfile.zip')

## Geography lookups

def lsoa_lookup(download=False):
    url = 'http://ons.maps.arcgis.com/sharing/rest/content/items/baf6566ca08949c6bbd61ff81d9526da/data'
    file_to_extract = 'OA11_LSOA11_MSOA11_LAD11_EW_LUv2.csv'
    save_as = 'ons-geography-lsoa-codes.csv'
    if download:
        download_unzip(url, file_to_extract, save_as)
    df = pd.read_csv('open-data/' + save_as, encoding = 'latin', dtype = str)
    df_formatted = df[['LSOA11CD', 'LSOA11NM', 'LAD11CD', 'LAD11NM']].rename(columns={'LSOA11CD': 'LSOA_code',
                                                                                  'LAD11CD': 'LA_Code'})
    return(df_formatted)

lsoa_lookup(False).head()

## Urban rural classification
The data is published as a machine readable csv

def clean_urban_rural_classification():
    df = pd.read_csv('open-data/urban-rural-classification.csv')
    df_clean = df[['LSOA11CD', 'RUC11']].rename(columns={'LSOA11CD': 'LSOA_code', 'RUC11': 'urban_rural'})
    return(df_clean)

def urban_rural_classification(download=False):
    url='http://ons.maps.arcgis.com/sharing/rest/content/items/9855221596994bde8363a685cb3dd58a/data'
    filepath='open-data/urban-rural-classification.csv'
    if download:
        download_file(url, filepath)
    df = pd.read_csv(filepath)
    df_clean = df[['LSOA11CD', 'RUC11']].rename(columns={'LSOA11CD': 'LSOA_code', 'RUC11': 'urban_rural'})
    return(df_clean)


urban_rural_classification(download=False).head()

## Schools

def download_schools_data():
    urls = {'2015': 'https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/445755/SFR16_2015_LA_tables.xlsx',
            '2014': 'https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/406915/SFR15_2014_LA_tables_v102.xlsx',
            '2016': 'https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/533028/SFR20_2016_LA_Tables_2016.xlsx'}

    for year, url in urls.items():
        filepath = 'open-data/schools-' + year + '.xlsx'
        download_file(url, filepath)

def extract_school_data(filepath):
    sheet = pyexcel.get_sheet(file_name=filepath,
                                sheet_name='Table 7a', 
                             start_row = 12)

    df = pd.DataFrame({
        'LA_Code': sheet.column[1], 
        'schools_nursery_LA': sheet.column[6],
        'schools_primary_LA': sheet.column[8],
        'schools_secondary_LA': sheet.column[10],
        'schools_private_LA': sheet.column[18],
        'schools_all_LA': sheet.column[20],
    })
    df_clean = df[df.LA_Code != '']
    return(df_clean)

def extract_school_data_2016(filepath):
    """The 2016 file is different to the 2014 and 2015 one"""
    sheet = pyexcel.get_sheet(file_name=filepath,
                                sheet_name='Table 7a', 
                             start_row = 12)

    df = pd.DataFrame({
        'LA_Code': sheet.column[1], 
        'schools_nursery_LA': sheet.column[4],
        'schools_primary_LA': sheet.column[6],
        'schools_secondary_LA': sheet.column[8],
        'schools_private_LA': sheet.column[16],
        'schools_all_LA': sheet.column[18],
    })
    df_clean = df[df.LA_Code != '']
    return(df_clean)

def school_data(download=False):
    if download:
        download_schools_data()
    df_14 = extract_school_data('open-data/schools-2014.xlsx')
    df_15 = extract_school_data('open-data/schools-2015.xlsx')
    df_16 = extract_school_data_2016('open-data/schools-2016.xlsx')
    df_14['year'] = 2014
    df_15['year'] = 2015
    df_16['year'] = 2016
    
    df = pd.concat([df_14, df_15, df_16], axis = 0, ignore_index=True).sort_values(['year', 'LA_Code']).reset_index(drop=True)
    # Some LA codes are incorrect and actually for counties. I created a lookup file to fix this
    la_code_fix = pd.read_csv('correct-la-codes.csv')
    df = pd.merge(df, la_code_fix, how='left', on='LA_Code')
    df1 = df[df.correct_la_code.isnull()].drop('correct_la_code', axis=1)
    df2 = df[df.correct_la_code.isnull()==False].drop('LA_Code', axis=1).rename(columns={'correct_la_code': 'LA_Code'})
    df = pd.concat([df1, df2], axis=0, ignore_index=True).sort_values(['year', 'LA_Code']).reset_index(drop=True)
    # Expand each LA to have a record for each of its LSOAs
    df_lsoa = pd.merge(df, lsoa_lookup(), how='left', on='LA_Code')
    df_lsoa = df_lsoa.drop(['LSOA11NM', 'LAD11NM', 'LA_Code'], axis=1).drop_duplicates()
    # Force to numeric
    for s in ['nursery', 'primary', 'secondary', 'private', 'all']:
        colname = 'schools_' + s + '_LA'
        df_lsoa[colname] = pd.to_numeric(df_lsoa[colname]) 
    return(df_lsoa)
    


school_data(download=False).head()

## Road lengths
The latest file published by DfT is a zip file. Inside it is a time series dating back to early 2000s

def road_length(download=False):
    
    if download:
        url = 'https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/610542/road-lengths-in-great-britain-2016.zip'
        file_to_extract = 'RDL0202 Total road length (kilometres) by road type and local authority in Great Britain, 2005 to 2016.ods'
        save_as = 'road-lengths.ods'
        download_unzip(url, file_to_extract, save_as)
        
    def extractor(year, last_four_columns_indicies):
        filepath = 'open-data/road-lengths.ods'
        sheet = pyexcel.get_sheet(file_name=filepath,
                                  sheet_name='RDL0202a_(' + str(year) + ')', 
                                  start_row = 7,
                                  row_limit = 206)
        df = pd.DataFrame({
            'LA_Code': sheet.column[0],
            'road_LA_trunk_length_km': sheet.column[last_four_columns_indicies[0]],
            'road_LA_principal_length_km': sheet.column[last_four_columns_indicies[1]],
            'road_LA_minor_length_km': sheet.column[last_four_columns_indicies[2]],
            'road_LA_total_length_km': sheet.column[last_four_columns_indicies[3]],
            'year': int(year)
        })
        return(df)
    
    df_2014 = extractor(2014, [20, 21, 22, 23] )
    df_2015 = extractor(2015, [22, 23, 24, 25] )
    df_2016 = extractor(2016, [20, 21, 22, 23] )
    df = pd.concat([df_2014, df_2015, df_2016], axis=0, ignore_index=True).sort_values(['year', 'LA_Code']).reset_index(drop=True)
    # Some LA codes are incorrect and actually for counties. I created a lookup file to fix this
    la_code_fix = pd.read_csv('correct-la-codes.csv')
    df = pd.merge(df, la_code_fix, how='left', on='LA_Code')
    df1 = df[df.correct_la_code.isnull()].drop('correct_la_code', axis=1)
    df2 = df[df.correct_la_code.isnull()==False].drop('LA_Code', axis=1).rename(columns={'correct_la_code': 'LA_Code'})
    df = pd.concat([df1, df2], axis=0, ignore_index=True).sort_values(['year', 'LA_Code']).reset_index(drop=True)
    # Expand each LA to have a record for each of its LSOAs
    df_lsoa = pd.merge(df, lsoa_lookup(), how='left', on='LA_Code')
    df_lsoa = df_lsoa.drop(['LSOA11NM', 'LAD11NM', 'LA_Code'], axis=1).drop_duplicates()
    return(df_lsoa)

road_length(download=False).head()

## Bus statistics
Look at how far buses have travelled a year by LA. Old data was found on national archives

def bus_travel(download=False):
    
    def download_bus_travel_stats():
        urls = {
            '2014': 'http://webarchive.nationalarchives.gov.uk/20150503071437/https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/413089/bus0208.xls',
            '2015': 'http://webarchive.nationalarchives.gov.uk/20160512013009/https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/463809/bus0208.xls',
            '2016': 'https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/560582/bus0208.ods'
        }
        for year, url in urls.items():
            download_file(url, 'open-data/bus-travel-' + year + '.' + url[-3:] )

    def extractor(filename):
        year = int(filename[21:25])
        sheet = pyexcel.get_sheet(file_name=filename,
                                          sheet_name='BUS0208b', 
                                          start_row = 8,
                                          row_limit = 120)
        df = pd.DataFrame({
            'LA_Code': sheet.column[0],
            'LA_name': sheet.column[2],
            'bus_LA_vehicle_km_travelled': sheet.column[6],
            'year': year
        })
        df = df[df.LA_name != ''].drop('LA_name', axis=1)
        return(df)
    
    if download:
        download_bus_travel_stats()
    df_2014 = extractor('open-data/bus-travel-2014.xls')
    df_2015 = extractor('open-data/bus-travel-2015.xls')
    df_2016 = extractor('open-data/bus-travel-2016.ods')
    df = pd.concat([df_2014, df_2015, df_2016], axis=0, ignore_index=True).sort_values(['year', 'LA_Code']).reset_index(drop=True)
    # Some LA codes are incorrect and actually for counties. I created a lookup file to fix this
    la_code_fix = pd.read_csv('correct-la-codes.csv')
    # Manual fix Tyne and Wear
    df.loc[df.LA_Code == 'E11000004', 'LA_Code'] = 'E11000007'
    df = pd.merge(df, la_code_fix, how='left', on='LA_Code')
    df1 = df[df.correct_la_code.isnull()].drop('correct_la_code', axis=1)
    df2 = df[df.correct_la_code.isnull()==False].drop('LA_Code', axis=1).rename(columns={'correct_la_code': 'LA_Code'})
    df = pd.concat([df1, df2], axis=0, ignore_index=True).sort_values(['year', 'LA_Code']).reset_index(drop=True)
    # Expand each LA to have a record for each of its LSOAs
    df_lsoa = pd.merge(df, lsoa_lookup(), how='left', on='LA_Code')
    df_lsoa = df_lsoa.drop(['LSOA11NM', 'LAD11NM', 'LA_Code'], axis=1).drop_duplicates()
    # Force to numeric
    df_lsoa['bus_LA_vehicle_km_travelled'] = pd.to_numeric(df_lsoa.bus_LA_vehicle_km_travelled)
    return(df_lsoa)

bus_travel(download=False).head()

## LSOA area size
Taken from a population density table

def area_km(download=False):
    if download:
        url = 'https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/lowersuperoutputareapopulationdensity/mid2015sape18dt11/sape18dt11mid2015lsoapopulationdensity.zip'
        file_to_extract = 'SAPE18DT11-mid-2015-lsoa-population-density.xls'
        save_as = 'population-density.xls'
        download_unzip(url, file_to_extract, save_as)
        
    sheet = pyexcel.get_sheet(file_name='open-data/population-density.xls',
                                              sheet_name='Mid-2015 Density', 
                                              start_row = 1,
                                              row_limit = 34755)
    df = pd.DataFrame({
        'LSOA_code': sheet.column[0],
        'area_square_km': sheet.column[3]
    })
    df = df.sort_values('LSOA_code').reset_index(drop=True)
    return(df)

area_km().head()

## LSOA population

def lsoa_population(download=False):
    def extractor(filename, sheet_name, year):
        d = pyexcel.get_sheet(file_name=filename, 
                              sheet_name=sheet_name,
                              name_columns_by_row=0) 
        df = pd.DataFrame(d.to_dict())
        df['year'] = int(year)
        return(df)
    
    params = [
        {
            'year': 2014,
            'url': 'https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/lowersuperoutputareamidyearpopulationestimates/mid2014/rft-lsoa-unformatted-table-2014.zip',
            'file_to_extract':'SAPE17DT2-mid-2014-lsoa-syoa-estimates-unformatted.xls',
            'save_as': 'population-2014.xls',
            'sheet': 'Mid-2014 Persons'},
        {
            'year': 2015,
            'url': 'https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/populationandmigration/populationestimates/datasets/lowersuperoutputareamidyearpopulationestimates/mid2015sape18dt1/sape18dt2mid2015lsoasyoaestimatesunformatted.zip',
            'file_to_extract': 'SAPE18DT2-mid-2015-lsoa-syoa-estimates-unformatted.xls',
            'save_as': 'population-2015.xls',
            'sheet': 'Mid-2015 Persons'
        }
    ]

    dfs = []
    for item in params:
        if download:
            download_unzip(item['url'], item['file_to_extract'], item['save_as'])
        dfs.append(extractor('open-data/' + item['save_as'], item['sheet'], item['year']))
    df = pd.concat(dfs, axis=0, ignore_index=True).rename(columns={'Code': 'LSOA_code'}).drop(['Name', 'All Ages'], axis=1)
    df_long = pd.melt(df, id_vars = ['LSOA_code', 'year'], var_name='age', value_name='population')
    df_long['age'] = pd.to_numeric(df_long.age)
    df_long['pop_u18'] = (df_long.age < 18) * df_long.population
    df_long['pop_working_age'] = ((df_long.age >= 18) & (df_long.age < 65 )) * df_long.population
    df_long['pop_retired_age'] = (df_long.age >= 65) * df_long.population
    df_long.drop(['age', 'population'], axis=1, inplace=True)
    df_aggregated = df_long.groupby(['LSOA_code', 'year']).sum().reset_index()

    return(df_aggregated)


lsoa_population(False).head()

# Putting it all together

def enriched_data():
    # Printing checks that now extra rows get added and observe how many new columns are added
    df = pd.read_csv('01_clean_stats.csv')
    print(df.shape)
    df = df.merge(urban_rural_classification(), how='left', on='LSOA_code')
    print(df.shape)
    df = df.merge(area_km(), how='left', on='LSOA_code')
    print(df.shape)
    df = df.merge(lsoa_population(), how='left', on=['LSOA_code', 'year'])
    print(df.shape)
    df = df.merge(road_length(), how='left', on=['LSOA_code', 'year'])
    print(df.shape)
    df = df.merge(bus_travel(), how='left', on=['LSOA_code', 'year'])
    print(df.shape)
    df = df.merge(school_data(), how='left', on=['LSOA_code', 'year'])
    print(df.shape)
    
    # London bus data is not recorded at LA level. Take the Tfl total and divide by 33 to split evenly accross LAs
    london_2016 = 487.9 / 33
    london_2015 = 485.0 /33
    london_2014 = 486.1 / 33
    f_2014 = (df.bus_LA_vehicle_km_travelled.isnull()) & (df.year == 2014)
    f_2015 = (df.bus_LA_vehicle_km_travelled.isnull()) & (df.year == 2015)
    f_2016 = (df.bus_LA_vehicle_km_travelled.isnull()) & (df.year == 2016)
    df.loc[f_2014, 'bus_LA_vehicle_km_travelled'] = london_2014
    df.loc[f_2015, 'bus_LA_vehicle_km_travelled'] = london_2015
    df.loc[f_2016, 'bus_LA_vehicle_km_travelled'] = london_2016

    return(df)

df = enriched_data()
df.to_csv('02-enriched-data.csv', index=False)

df.head()

df.describe()
