==> 08-school-postcodes.ipynb <==
School data is available on the [DfE](https://www.compare-school-performance.service.gov.uk/) website that includes the postcode for each school.
We can convert a postcode to lat-lon using the free api [Postcodes.io](https://postcodes.io/)
We can then use lat-lon of the postcodes and the LSOA midpoint to measure the closest distance to a school and add it  as variavble in the machine learning. Hopefully greatly improving model accuracy.

The API would take a while to install. It uses data from ONS available online [here](https://data.gov.uk/dataset/national-statistics-postcode-lookup-uk)

## Download DfE data

import pandas as pd
import numpy as np


url_2015 = 'https://www.compare-school-performance.service.gov.uk/download-data?download=true&regions=0&filters=SPINE&fileformat=csv&year=2014-2015&meta=false'
url_2014 = 'https://www.compare-school-performance.service.gov.uk/download-data?download=true&regions=0&filters=SPINE&fileformat=csv&year=2013-2014&meta=false'

save_as_2015 = 'open-data/schools-v2-2015.csv'
save_as_2014 = 'open-data/schools-v2-2014.csv'

Cannot download in python. Do it manually

## Download postcode data

Manually download the csv from [https://data.gov.uk/dataset/national-statistics-postcode-lookup-uk](https://data.gov.uk/dataset/national-statistics-postcode-lookup-uk) and save as 'open-data/postcodes.csv

df_postcodes = pd.read_csv('open-data/postcodes.csv')

df_postcodes['Postcode 1'] = df_postcodes['Postcode 1'].apply(lambda x: x.upper().replace(' ', ''))
df_postcodes['Postcode 2'] = df_postcodes['Postcode 1'].apply(lambda x: x.upper().replace(' ', ''))
df_postcodes['Postcode 3'] = df_postcodes['Postcode 1'].apply(lambda x: x.upper().replace(' ', ''))
p1 = df_postcodes.drop(['Postcode 2', 'Postcode 3'], axis=1).rename(columns={'Postcode 1': 'POSTCODE'})
p2 = df_postcodes.drop(['Postcode 1', 'Postcode 3'], axis=1).rename(columns={'Postcode 2': 'POSTCODE'})
p3 = df_postcodes.drop(['Postcode 1', 'Postcode 2'], axis=1).rename(columns={'Postcode 3': 'POSTCODE'})
df_postcodes_long = pd.concat([p1, p2, p3], axis = 0, ignore_index=True)


df_postcodes_long = df_postcodes_long[['POSTCODE','Lower Super Output Area Code',
                            'Longitude', 'Latitude']].drop_duplicates()
df_postcodes_long = df_postcodes_long[df_postcodes_long['Lower Super Output Area Code'].isnull() == False]

df_postcodes_long.head()

df_postcodes_long.shape

### Merge the two datasets on the postcodes

def school_by_lsoa(year):
    if year == 2014:
        file = save_as_2014
    if year == 2015:
        file = save_as_2015
    
    df_schools = pd.read_csv(file)
    df_schools = df_schools[['URN', 'SCHNAME', 'POSTCODE', 'ISPRIMARY', 'ISSECONDARY', 'ISPOST16']]
    df_schools = df_schools[df_schools.POSTCODE.isnull() == False]
    df_schools['POSTCODE'] = df_schools['POSTCODE'].apply(lambda x: x.upper().replace(' ', ''))
    merged = df_schools.merge(df_postcodes_long, how='left', on='POSTCODE')
    not_merged = merged['Lower Super Output Area Code'].isnull().sum() / len(merged)
    print('Percent of schools were the LSOA could not be determined from the postcode')
    print('{:.2f}'.format(not_merged*100) + '%')
    merged['year'] = year
    return(merged)

df_2014 = school_by_lsoa(2014)
df_2015 = school_by_lsoa(2015)
df = pd.concat([df_2014, df_2015], ignore_index = False).rename(columns={'Lower Super Output Area Code': 'LSOA_code'})
secondary = df[df.ISSECONDARY == 1]
primary = df[df.ISPRIMARY == 1]
post16 = df[df.ISPOST16 == 1]



secondary = secondary.groupby(['year', 'LSOA_code']).size().reset_index(name='schools_secondary')
primary = primary.groupby(['year', 'LSOA_code']).size().reset_index(name='schools_primary')
post16 = post16.groupby(['year', 'LSOA_code']).size().reset_index(name='schools_post16')

df = primary.merge(secondary).merge(post16)

## Codes are 2001. Convert to 2011
Download Data
http://geoportal.statistics.gov.uk/datasets?q=Lower%20Layer%20Super%20Output%20Area%20(2001)%20to%20Lower%20Layer%20Super%20Output%20Area%20(2011)%20to%20Local%20Authority%20District%20(2011)%20Lookup&sort=name

lsoa_lookup = pd.read_csv('open-data/LSOA01_to_LSOA11.csv', encoding = 'latin')

lsoa_lookup.sort_values('LSOA11CD').head()

df = df.merge(lsoa_lookup, how='left', left_on='LSOA_code', right_on='LSOA01CD')

df_2011_zones = df.groupby(['year', 'LSOA11CD' ])[['schools_primary', 'schools_secondary', 'schools_post16' ]].sum().reset_index()
df_2011_zones = df_2011_zones.rename(columns={'LSOA11CD': 'LSOA_code'})

df_2011_zones.head()

## Join the new data onto the existing lot

df_old = pd.read_csv('02-enriched-data.csv')

df_new = df_old.merge(df_2011_zones, how='left', on=['LSOA_code', 'year'])

df_new.schools_secondary.isnull().sum() / len(df_new)

city_of_london_lsoa = 'E01000001'
city_of_london_lsoa in df.LSOA_code

df_new[df_new.schools_primary.isnull() == False]

df_new.head()
